backend/requirements.txt
aiofiles==0.4.0
asyncio==3.4.3
click==7.1.2
h11==0.8.1
httptools==0.0.13
python-decouple==3.3
python-multipart==0.0.5
six==1.15.0
starlette==0.12.0
uvicorn==0.7.1
uvloop==0.12.2
websockets==7.0
